project running

+++ app/src/main/kotlin/com/chaitanya/easyrelocate/model/Location.kt
@@ -1,7 +1,7 @@
 package com.chaitanya.easyrelocate.model
 
 data class Location (
-       var lat :  Int,
-       var lng :  Int,
+       var lat :  Double,
+       var lng :  Double,
        var address: String
 )
